runner_service/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
gitdb==4.0.12
gitpython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.1
typing-extensions==4.13.1
typing-inspection==0.4.0
uvicorn==0.34.0
